# src/SBD.ContentInsights/Models/DocumentType.cs
namespace SBD.ContentInsights.Models;

public class DocumentType
{
    public string Type { get; set; } = string.Empty;
    public string Name { get; set; } = string.Empty;
    public int Count { get; set; }
}