# src/SBD.ContentInsights/client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Content Insights",
    "version": "1.0"
  },
  "paths": {
    "/umbraco/management/api/v1/content-insights/get-all-documents-with-authors": {
      "get": {
        "tags": [
          "v1"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DocumentsWithAuthors"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          },
          "403": {
            "description": "The authenticated user does not have access to this resource"
          }
        },
        "security": [
          {
            "Backoffice User": [ ]
          }
        ]
      }
    },
    "/umbraco/management/api/v1/content-insights/get-document-types": {
      "get": {
        "tags": [
          "v1"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/DocumentType"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "The resource is protected and requires an authentication token"
          },
          "403": {
            "description": "The authenticated user does not have access to this resource"
          }
        },
        "security": [
          {
            "Backoffice User": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Author": {
        "required": [
          "email",
          "link",
          "name",
          "userGroups"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "userGroups": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/UserGroup"
                }
              ]
            }
          }
        },
        "additionalProperties": false
      },
      "DocumentType": {
        "required": [
          "count",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DocumentsWithAuthors": {
        "required": [
          "authors",
          "documents"
        ],
        "type": "object",
        "properties": {
          "documents": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/UmbracoDocument"
                }
              ]
            }
          },
          "authors": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Author"
                }
              ]
            }
          }
        },
        "additionalProperties": false
      },
      "UmbracoDocument": {
        "required": [
          "authorKey",
          "link",
          "name",
          "status",
          "type",
          "typeName",
          "updateDate"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "typeName": {
            "type": "string"
          },
          "authorKey": {
            "type": "string"
          },
          "updateDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UserGroup": {
        "required": [
          "key",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Backoffice User": {
        "type": "oauth2",
        "description": "Umbraco Authentication",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/umbraco/management/api/v1/security/back-office/authorize",
            "tokenUrl": "/umbraco/management/api/v1/security/back-office/token",
            "scopes": { }
          }
        }
      }
    }
  }
}